Changed: - refactored apply button to apply_button.dart
