+++ lib/pages/details.dart
@@ -1,9 +1,9 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:thewanted/pages/status_tag.dart';
 import 'package:thewanted/models/user.dart';
 import 'package:thewanted/models/skills.dart';
+import 'package:thewanted/pages/details_components/apply_button.dart';
 
 class Request {
   //String userName;
@@ -318,101 +318,11 @@ class _DetailedPageState extends State<DetailedPage> {
                 child: labels,
               ),
               contactInfo,
-              _buildApplyButton(),
+              ApplyButton(taskId: widget.id, status: this.request.status),
             ],
           ),
         ));
   }
-
-  Widget _buildApplyButton() => Padding(
-        padding: const EdgeInsets.symmetric(vertical: 16.0),
-        child: RaisedButton(
-          color: _buildColorFromStatus(),
-          onPressed: _buildOnpressedFromStatus(),
-          child: _buildTextFromStatus(),
-        ),
-      );
-
-  Text _buildTextFromStatus() {
-    switch (StatusTag.getStatusFromString(this.request.status)) {
-      case Status.open:
-        return Text('Apply');
-        break;
-      case Status.inprogress:
-        return Text('In Progress');
-        break;
-      case Status.finished:
-        return Text('Closed');
-        break;
-      default:
-        return Text('Undefined');
-        break;
-    }
-  }
-
-  Color _buildColorFromStatus() {
-    switch (StatusTag.getStatusFromString(this.request.status)) {
-      case Status.open:
-        return Colors.green;
-        break;
-      case Status.inprogress:
-        return Colors.yellow[800];
-        break;
-      case Status.finished:
-        return Colors.red;
-        break;
-      default:
-        return Colors.blue;
-        break;
-    }
-  }
-
-  VoidCallback _buildOnpressedFromStatus() {
-    switch (StatusTag.getStatusFromString(this.request.status)) {
-      case Status.open:
-        return _apply;
-        break;
-      default:
-        return null;
-        break;
-    }
-  }
-
-  _apply() {
-    print('Apply this task!');
-    // TODO: implement apply
-    // 1. ask user to input apply message
-    // 2. add apply to firestore
-    _updateTaskdataAndProfiledata();
-    // 3. Send owner (email) notification
-  }
-
-  Future<void> _updateTaskdataAndProfiledata() =>
-      _getCurrentUserId().then((FirebaseUser user) {
-        _updateTaskApplicants(user.uid);
-        _updateProfileApplied(user.uid);
-      }).catchError((e) {
-        print(e);
-      });
-
-  Future<void> _updateTaskApplicants(String uid) => Firestore.instance
-          .collection('tasks')
-          .document(widget.id)
-          .collection('applicants')
-          .document(uid)
-          .setData({
-        'des': 'sdfasdfaefasd',
-      });
-
-  Future<void> _updateProfileApplied(String uid) =>
-      Firestore.instance.collection('users').document(uid).updateData({
-        'applied': FieldValue.arrayUnion([
-          widget.id,
-        ])
-      });
-
-  Future<FirebaseUser> _getCurrentUserId() =>
-      FirebaseAuth.instance.currentUser();
 }
 
 class userProfileInfoPage extends StatefulWidget {
